completed program

+++ functions.h
@@ -2,6 +2,6 @@
 int addition(int a, int b);
 int subtraction(int a, int b);
 int multiplication(int a, int b);
-double division(int a, int b);
+double division(double a, double b);
 void showMenu();
